+++ project/src/components/organisms/LoginForm/LoginForm.tsx
@@ -1,0 +1,56 @@
+import React from 'react';
+import styled from 'styled-components';
+import StyledH1 from '../../atoms/Typography/StyledH1';
+import allColors from '../../../constants/allColors';
+import LoginInput from './LoginInput';
+
+const LoginWrapper = styled.div`
+    height: 66vh;
+    width: 25vw;
+    background-color: white;
+    border-radius: 2px;
+    padding: 20px;
+    box-shadow: 5px 5px 20px -8px rgba(0, 0, 0, 1);
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+`;
+
+const StyledParagraph = styled.div`
+    font-family: 'Montserrat', sans-serif;
+    font-size: 12px;
+    font-weight: 700;
+    color: ${allColors.darkGrey};
+    margin-left: 5px;
+    margin-top: 10px;
+`;
+
+const BottomWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 70%;
+    width: 100%;
+`;
+
+const LoginForm = () => {
+    return (
+        <LoginWrapper>
+            <div>
+                <StyledH1 size={28} color={allColors.lightBlue}>
+                    {'Welcome'}
+                </StyledH1>
+                <StyledParagraph>Please Login to continue</StyledParagraph>
+            </div>
+            <BottomWrapper>
+                <LoginInput
+                    id="emailInput"
+                    placeholderText="Type your email..."
+                    iconPath={require('../../../assets/SVGs/email.svg')}
+                />
+            </BottomWrapper>
+        </LoginWrapper>
+    );
+};
+
+export default LoginForm;